+++ src/NotificationsPro/Converters/TimestampTextConverter.cs
@@ -1,5 +1,6 @@
 using System.Globalization;
 using System.Windows.Data;
+using NotificationsPro.Helpers;
 
 namespace NotificationsPro.Converters;
 
@@ -13,14 +14,7 @@ public class TimestampTextConverter : IMultiValueConverter
         var relative = values.Length > 0 ? values[0]?.ToString() : string.Empty;
         var receivedAt = values.Length > 1 && values[1] is DateTime dt ? dt : DateTime.Now;
         var mode = values.Length > 2 ? values[2]?.ToString() : "Relative";
-
-        if (string.Equals(mode, "Time", StringComparison.OrdinalIgnoreCase))
-            return receivedAt.ToString("HH:mm", culture);
-
-        if (string.Equals(mode, "DateTime", StringComparison.OrdinalIgnoreCase))
-            return receivedAt.ToString("g", culture);
-
-        return string.IsNullOrWhiteSpace(relative) ? "just now" : relative;
+        return TimestampTextFormatter.Format(relative, receivedAt, mode, culture);
     }
 
     public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)